Rename rest package to server
