+++ tests/rest_helpers_test.go
@@ -7,14 +7,14 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/snickers/snickers/rest"
+	"github.com/snickers/snickers/server"
 )
 
-var _ = Describe("Rest API", func() {
+var _ = Describe("Server API", func() {
 	Context("helper functions", func() {
 		It("should write the error as json", func() {
 			w := httptest.NewRecorder()
-			rest.HTTPError(w, http.StatusOK, "error here", errors.New("database broken"))
+			server.HTTPError(w, http.StatusOK, "error here", errors.New("database broken"))
 
 			Expect(w.Code).To(Equal(http.StatusOK))
 			Expect(w.Body.String()).To(Equal(`{"error": "error here: database broken"}`))
