+++ tests/rest_job_test.go
@@ -11,20 +11,20 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/snickers/snickers/db"
-	"github.com/snickers/snickers/rest"
+	"github.com/snickers/snickers/server"
 	"github.com/snickers/snickers/types"
 )
 
-var _ = Describe("Rest API", func() {
+var _ = Describe("Server API", func() {
 	var (
 		response   *httptest.ResponseRecorder
-		server     *mux.Router
+		s          *mux.Router
 		dbInstance db.DatabaseInterface
 	)
 
 	BeforeEach(func() {
 		response = httptest.NewRecorder()
-		server = rest.NewRouter()
+		s = server.NewRouter()
 		dbInstance, _ = db.GetDatabase()
 		dbInstance.ClearDatabase()
 	})
@@ -32,7 +32,7 @@ var _ = Describe("Rest API", func() {
 	Describe("GET /jobs", func() {
 		It("should return application/json on its content type", func() {
 			request, _ := http.NewRequest("GET", "/jobs", nil)
-			server.ServeHTTP(response, request)
+			s.ServeHTTP(response, request)
 			Expect(response.HeaderMap["Content-Type"][0]).To(Equal("application/json; charset=UTF-8"))
 		})
 
@@ -46,7 +46,7 @@ var _ = Describe("Rest API", func() {
 			expected2, _ := json.Marshal(`[{"id":"321","source":"","destination":"","preset":{"video":{},"audio":{}},"status":"","progress":""},{"id":"123","source":"","destination":"","preset":{"video":{},"audio":{}},"status":"","progress":""}]`)
 
 			request, _ := http.NewRequest("GET", "/jobs", nil)
-			server.ServeHTTP(response, request)
+			s.ServeHTTP(response, request)
 			responseBody, _ := json.Marshal(response.Body.String())
 
 			Expect(response.Code).To(Equal(http.StatusOK))
@@ -59,7 +59,7 @@ var _ = Describe("Rest API", func() {
 			dbInstance.StorePreset(types.Preset{Name: "presetName"})
 			jobJSON := []byte(`{"source": "http://flv.io/src.mp4", "destination": "s3://l@p:google.com", "preset": "presetName"}`)
 			request, _ := http.NewRequest("POST", "/jobs", bytes.NewBuffer(jobJSON))
-			server.ServeHTTP(response, request)
+			s.ServeHTTP(response, request)
 			responseBody := response.Body.String()
 
 			jobs, _ := dbInstance.GetJobs()
@@ -80,7 +80,7 @@ var _ = Describe("Rest API", func() {
 			It("should return BadRequest if preset is not set", func() {
 				jobJSON := []byte(`{"source": "http://flv.io/src.mp4", "destination": "s3://l@p:google.com", "preset": "presetName"}`)
 				request, _ := http.NewRequest("POST", "/jobs", bytes.NewBuffer(jobJSON))
-				server.ServeHTTP(response, request)
+				s.ServeHTTP(response, request)
 				responseBody, _ := json.Marshal(response.Body.String())
 
 				expected, _ := json.Marshal(`{"error": "retrieving preset: preset not found"}`)
@@ -92,7 +92,7 @@ var _ = Describe("Rest API", func() {
 			It("should return BadRequest if preset is malformed", func() {
 				jobJSON := []byte(`{"source: "http://flv.io/src.mp4", "destinat}`)
 				request, _ := http.NewRequest("POST", "/jobs", bytes.NewBuffer(jobJSON))
-				server.ServeHTTP(response, request)
+				s.ServeHTTP(response, request)
 				responseBody, _ := json.Marshal(response.Body.String())
 
 				expected, _ := json.Marshal(`{"error": "unpacking job: invalid character 'h' after object key"}`)
@@ -117,7 +117,7 @@ var _ = Describe("Rest API", func() {
 			expected, _ := json.Marshal(&job)
 
 			request, _ := http.NewRequest("GET", "/jobs/123-123-123", nil)
-			server.ServeHTTP(response, request)
+			s.ServeHTTP(response, request)
 			Expect(response.Body.String()).To(Equal(string(expected)))
 			Expect(response.Code).To(Equal(http.StatusOK))
 			Expect(response.HeaderMap["Content-Type"][0]).To(Equal("application/json; charset=UTF-8"))
@@ -126,7 +126,7 @@ var _ = Describe("Rest API", func() {
 		Context("when getting the job fails", func() {
 			It("should return BadRequest", func() {
 				request, _ := http.NewRequest("GET", "/jobs/123-456-9292", nil)
-				server.ServeHTTP(response, request)
+				s.ServeHTTP(response, request)
 				expected, _ := json.Marshal(`{"error": "retrieving job: job not found"}`)
 				responseBody, _ := json.Marshal(response.Body.String())
 				Expect(responseBody).To(Equal(expected))
@@ -149,7 +149,7 @@ var _ = Describe("Rest API", func() {
 			dbInstance.StoreJob(job)
 
 			request, _ := http.NewRequest("POST", "/jobs/123-123-123/start", nil)
-			server.ServeHTTP(response, request)
+			s.ServeHTTP(response, request)
 			Expect(response.Code).To(Equal(http.StatusOK))
 			Expect(response.HeaderMap["Content-Type"][0]).To(Equal("application/json; charset=UTF-8"))
 		})
@@ -157,7 +157,7 @@ var _ = Describe("Rest API", func() {
 		Context("when starting a job fails", func() {
 			It("should return BadRequest", func() {
 				request, _ := http.NewRequest("POST", "/jobs/123-456-9292/start", nil)
-				server.ServeHTTP(response, request)
+				s.ServeHTTP(response, request)
 				expected, _ := json.Marshal(`{"error": "retrieving job: job not found"}`)
 				responseBody, _ := json.Marshal(response.Body.String())
 				Expect(responseBody).To(Equal(expected))
